Update Project2_Final.Rmd
Here is my part without the write up (will do that later today). Everything should work except for maybe the wordclouds because they are in a function that takes in a column number of the df data frame. I will work on my explanations and fixing anything that doesn't work.